# openapi/beecount-cloud-v1.yaml
openapi: 3.1.0
info:
  title: BeeCount Cloud
  description: BeeCount Cloud v1 API
  version: 0.1.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/sync/push:
    post:
      tags:
      - sync
      summary: Push Changes
      operationId: push_changes_api_v1_sync_push_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
...
  /api/v1/attachments/upload:
    post:
      tags:
      - attachments
      summary: Upload Attachment
      operationId: upload_attachment_api_v1_attachments_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_attachment_api_v1_attachments_upload_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentUploadOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - app_write
  /api/v1/attachments/batch-exists:
    post:
      tags:
      - attachments
      summary: Batch Exists
      operationId: batch_exists_api_v1_attachments_batch_exists_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachmentBatchExistsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentBatchExistsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - app_write
  /api/v1/attachments/{file_id}:
    get:
      tags:
      - attachments
      summary: Download Attachment
      operationId: download_attachment_api_v1_attachments__file_id__get
      security:
      - OAuth2PasswordBearer:
        - app_write
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AttachmentBatchExistsRequest:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        sha256_list:
          items:
            type: string
          type: array
          title: Sha256 List
      type: object
      required:
      - ledger_id
      title: AttachmentBatchExistsRequest
    AttachmentBatchExistsResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AttachmentExistsItem'
          type: array
          title: Items
      type: object
      title: AttachmentBatchExistsResponse
    AttachmentExistsItem:
      properties:
        sha256:
          type: string
          title: Sha256
        exists:
          type: boolean
          title: Exists
        file_id:
          anyOf:
          - type: string
          - type: 'null'
          title: File Id
        size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Size
        mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Mime Type
      type: object
      required:
      - sha256
      - exists
      title: AttachmentExistsItem
    AttachmentUploadOut:
      properties:
        file_id:
          type: string
          title: File Id
        ledger_id:
          type: string
          title: Ledger Id
        sha256:
          type: string
          title: Sha256
        size:
          type: integer
          title: Size
        mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Mime Type
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - file_id
      - ledger_id
      - sha256
      - size
      - created_at
      title: AttachmentUploadOut
    AuthLoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        platform:
          anyOf:
          - type: string
          - type: 'null'
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        client_type:
          type: string
          enum:
          - app
          - web
          title: Client Type
          default: app
      type: object
      required:
      - email
      - password
      title: AuthLoginRequest
...
    AuthRegisterRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 6
          title: Password
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        platform:
          anyOf:
          - type: string
          - type: 'null'
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        client_type:
          type: string
          enum:
          - app
          - web
          title: Client Type
          default: app
      type: object
      required:
      - email
      - password
      title: AuthRegisterRequest
    AuthTokenResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        expires_in:
          type: integer
          title: Expires In
        device_id:
          type: string
          title: Device Id
        scopes:
          items:
            type: string
          type: array
          title: Scopes
      type: object
      required:
      - user
      - access_token
      - refresh_token
      - expires_in
      - device_id
      title: AuthTokenResponse
    Body_upload_attachment_api_v1_attachments_upload_post:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - ledger_id
      - file
      title: Body_upload_attachment_api_v1_attachments_upload_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SyncChangeIn:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        entity_type:
          type: string
          title: Entity Type
        entity_sync_id:
          type: string
          title: Entity Sync Id
        action:
          type: string
          enum:
          - upsert
          - delete
          title: Action
        payload:
          additionalProperties: true
          type: object
          title: Payload
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - ledger_id
      - entity_type
      - entity_sync_id
      - action
      - payload
      - updated_at
      title: SyncChangeIn
...
    SyncPushRequest:
      properties:
        device_id:
          type: string
          title: Device Id
        changes:
          items:
            $ref: '#/components/schemas/SyncChangeIn'
          type: array
          title: Changes
      type: object
      required:
      - device_id
      - changes
      title: SyncPushRequest
    SyncPushResponse:
      properties:
        accepted:
          type: integer
          title: Accepted
        rejected:
          type: integer
          title: Rejected
        conflict_count:
          type: integer
          title: Conflict Count
          default: 0
        conflict_samples:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Conflict Samples
        server_cursor:
          type: integer
          title: Server Cursor
        server_timestamp:
          type: string
          format: date-time
          title: Server Timestamp
      type: object
      required:
      - accepted
      - rejected
      - server_cursor
      - server_timestamp
      title: SyncPushResponse
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        is_admin:
          type: boolean
          title: Is Admin
          default: false
      type: object
      required:
      - id
      - email
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes:
            app_write: App client read/write scope
            web_read: Web console read scope
            web_write: Web console write scope
            ops_write: Operations and admin scope
          tokenUrl: /api/v1/auth/login